Refactor row validation into a new package and update tests
