+++ loan_info/row_shape_validations.go
@@ -1,8 +1,11 @@
-package validator
+package loan_info
 
-import "fmt"
+import (
+	"fmt"
+	"go-file-parsing/validator"
+)
 
-func isValidSize(ctx *RowValidatorContext, cols []string) (map[string]string, error) {
+func isValidSize(ctx *validator.RowValidatorContext, cols []string) (map[string]string, error) {
 	if len(cols) != ctx.Config.ExpectedColumns {
 		return nil, fmt.Errorf("expected %d columns, got %d", ctx.Config.ExpectedColumns, len(cols))
 	}
